tests/wpt/web-platform-tests/websockets/Create-Secure-verify-url-set-non-default-port.htm

<!DOCTYPE html>
<html>
<head>
    <title>W3C WebSocket API - Create Secure WebSocket - wsocket.url is set correctly - non default port</title>
    <script type="text/javascript" src="/resources/testharness.js"></script>
    <script type="text/javascript" src="/resources/testharnessreport.js"></script>
    <script type="text/javascript" src="websocket.sub.js"></script>
</head>
<body>
    <div id="log"></div>
    <script type="text/javascript">
        test(function () {
            var urlNonDefaultPort = "wss://" + __SERVER__NAME + ":" + __NEW__SECURE__PORT + "/" + __PATH;
            var wsocket = new WebSocket(urlNonDefaultPort);
            assert_equals(wsocket.url, urlNonDefaultPort, "wsocket.url is set correctly");
        }, "W3C WebSocket API - Create Secure WebSocket - wsocket.url should be set correctly")
    </script>
</body>
</html>
